correct UMAP
