+++ torchdr/__init__.py
@@ -28,8 +28,7 @@ from .affinity import (
     SinkhornAffinity,
     StudentAffinity,
     SymmetricEntropicAffinity,
-    UMAPAffinityIn,
-    UMAPAffinityOut,
+    UMAPAffinity,
     PHATEAffinity,
     PACMAPAffinity,
 )
@@ -78,8 +77,7 @@ __all__ = [
     "SymmetricEntropicAffinity",
     "SinkhornAffinity",
     "DoublyStochasticQuadraticAffinity",
-    "UMAPAffinityIn",
-    "UMAPAffinityOut",
+    "UMAPAffinity",
     "DRModule",
     "KMeans",
     "AffinityMatcher",
