+++ torchdr/affinity/unnormalized.py
@@ -7,32 +7,9 @@
 
 import torch
 from typing import Union, Optional
-from scipy.optimize import curve_fit
-import numpy as np
 
 from torchdr.affinity.base import UnnormalizedAffinity, UnnormalizedLogAffinity
 from torchdr.utils import LazyTensorType, compile_if_requested
-
-
-# from umap/umap/umap_.py
-def find_ab_params(spread, min_dist):
-    """Fit a, b params as in UMAP.
-
-    Fit (a, b) for the differentiable curve used in lower
-    dimensional fuzzy simplicial complex construction. We want the
-    smooth curve (from a pre-defined family with simple gradient) that
-    best matches an offset exponential decay.
-    """
-
-    def curve(x, a, b):
-        return 1.0 / (1.0 + a * x ** (2 * b))
-
-    xv = np.linspace(0, spread * 3, 300)
-    yv = np.zeros(xv.shape)
-    yv[xv < min_dist] = 1.0
-    yv[xv >= min_dist] = np.exp(-(xv[xv >= min_dist] - min_dist) / spread)
-    params, covar = curve_fit(curve, xv, yv)
-    return params[0].item(), params[1].item()
 
 
 class GaussianAffinity(UnnormalizedLogAffinity):
@@ -303,79 +280,3 @@ class ScalarProductAffinity(NegativeCostAffinity):
             compile=compile,
             _pre_processed=_pre_processed,
         )
-
-
-class UMAPAffinityOut(UnnormalizedLogAffinity):
-    r"""Compute the affinity used in embedding space in UMAP :cite:`mcinnes2018umap`.
-
-    Its :math:`(i,j)` coefficient is as follows:
-
-    .. math::
-        1 / \left(1 + a C_{ij}^{b} \right)
-
-    where parameters a and b are fitted to the spread and min_dist parameters.
-
-    Parameters
-    ----------
-    min_dist : float, optional
-        min_dist parameter from UMAP. Provides the minimum distance apart that
-        points are allowed to be.
-    spread : float, optional
-        spread parameter from UMAP.
-    a : float, optional
-        factor of the cost matrix.
-    b : float, optional
-        exponent of the cost matrix.
-    degrees_of_freedom : int, optional
-        Degrees of freedom for the Student-t distribution.
-    metric : str, optional
-        Metric to use for pairwise distances computation.
-    zero_diag : bool, optional
-        Whether to set the diagonal of the affinity matrix to zero.
-    device : str, optional
-        Device to use for computations.
-    backend : {"keops", "faiss", None}, optional
-        Which backend to use for handling sparsity and memory efficiency.
-        Default is None.
-    verbose : bool, optional
-        Verbosity. Default is False.
-    compile : bool, optional
-        Whether to compile the formula.
-    _pre_processed : bool, optional
-        If True, assumes inputs are already torch tensors on the correct device
-        and skips the `to_torch` conversion. Default is False.
-    """
-
-    def __init__(
-        self,
-        min_dist: float = 0.1,
-        spread: float = 1,
-        a: Optional[float] = None,
-        b: Optional[float] = None,
-        metric: str = "sqeuclidean",
-        zero_diag: bool = True,
-        device: str = "auto",
-        backend: Optional[str] = None,
-        verbose: bool = False,
-        compile: bool = False,
-        _pre_processed: bool = False,
-    ):
-        if a is None or b is None:
-            a, b = find_ab_params(spread, min_dist)
-        super().__init__(
-            metric=metric,
-            zero_diag=zero_diag,
-            device=device,
-            backend=backend,
-            verbose=verbose,
-            compile=compile,
-            _pre_processed=_pre_processed,
-        )
-        self.min_dist = min_dist
-        self.spread = spread
-        self._a = a
-        self._b = b
-
-    @compile_if_requested
-    def _log_affinity_formula(self, C: torch.Tensor):
-        return -(1 + self._a * C**self._b).log()
