+++ torchdr/neighbor_embedding/umap.py
@@ -7,15 +7,43 @@
 from typing import Dict, Optional, Union, Type
 import torch
 
-from torchdr.affinity import UMAPAffinityIn, UMAPAffinityOut
+from torchdr.affinity import UMAPAffinity
 from torchdr.neighbor_embedding.base import SampledNeighborEmbedding
-from torchdr.utils import cross_entropy_loss, sum_red
+from torchdr.utils import (
+    cross_entropy_loss,
+    sum_red,
+    symmetric_pairwise_distances_indices,
+)
+
+from scipy.optimize import curve_fit
+import numpy as np
+
+
+# from umap/umap/umap_.py
+def find_ab_params(spread, min_dist):
+    """Fit a, b params as in UMAP.
+
+    Fit (a, b) for the differentiable curve used in lower
+    dimensional fuzzy simplicial complex construction. We want the
+    smooth curve (from a pre-defined family with simple gradient) that
+    best matches an offset exponential decay.
+    """
+
+    def curve(x, a, b):
+        return 1.0 / (1.0 + a * x ** (2 * b))
+
+    xv = np.linspace(0, spread * 3, 300)
+    yv = np.zeros(xv.shape)
+    yv[xv < min_dist] = 1.0
+    yv[xv >= min_dist] = np.exp(-(xv[xv >= min_dist] - min_dist) / spread)
+    params, covar = curve_fit(curve, xv, yv)
+    return params[0].item(), params[1].item()
 
 
 class UMAP(SampledNeighborEmbedding):
     r"""UMAP introduced in :cite:`mcinnes2018umap` and further studied in :cite:`damrich2021umap`.
 
-    It uses a :class:`~torchdr.UMAPAffinityIn` as input
+    It uses a :class:`~torchdr.UMAPAffinity` as input
     affinity :math:`\mathbf{P}` and a :class:`~torchdr.UMAPAffinityOut` as output
     affinity :math:`\mathbf{Q}`.
 
@@ -130,15 +158,18 @@ class UMAP(SampledNeighborEmbedding):
         self.n_neighbors = n_neighbors
         self.min_dist = min_dist
         self.spread = spread
-        self.a = a
-        self.b = b
         self.metric_in = metric_in
         self.metric_out = metric_out
         self.max_iter_affinity = max_iter_affinity
         self.tol_affinity = tol_affinity
         self.sparsity = sparsity
 
-        affinity_in = UMAPAffinityIn(
+        if a is None or b is None:
+            a, b = find_ab_params(self.spread, self.min_dist)
+        self._a = a
+        self._b = b
+
+        affinity_in = UMAPAffinity(
             n_neighbors=n_neighbors,
             metric=metric_in,
             tol=tol_affinity,
@@ -147,28 +178,11 @@ class UMAP(SampledNeighborEmbedding):
             backend=backend,
             verbose=verbose,
             sparsity=sparsity,
+            compile=compile,
         )
-        affinity_out = UMAPAffinityOut(
-            min_dist=min_dist,
-            spread=spread,
-            a=a,
-            b=b,
-            metric=metric_out,
-            device=device,
-            verbose=False,
-        )
-
-        _scheduler_kwargs = scheduler_kwargs
-        if scheduler == "LinearLR" and scheduler_kwargs == "auto":
-            _scheduler_kwargs = {
-                "start_factor": 1.0,
-                "end_factor": 0,
-                "total_iters": max_iter,
-            }
 
         super().__init__(
             affinity_in=affinity_in,
-            affinity_out=affinity_out,
             n_components=n_components,
             optimizer=optimizer,
             optimizer_kwargs=optimizer_kwargs,
@@ -176,7 +190,7 @@ class UMAP(SampledNeighborEmbedding):
             max_iter=max_iter,
             lr=lr,
             scheduler=scheduler,
-            scheduler_kwargs=_scheduler_kwargs,
+            scheduler_kwargs=scheduler_kwargs,
             init=init,
             init_scaling=init_scaling,
             device=device,
@@ -191,11 +205,15 @@ class UMAP(SampledNeighborEmbedding):
         )
 
     def _repulsive_loss(self):
-        Q = self.affinity_out(self.embedding_, indices=self.neg_indices_)
-        Q = Q / (Q + 1)  # stabilization trick, PR #856 from UMAP repo
-        return -sum_red((1 - Q).log(), dim=(0, 1))
+        D = symmetric_pairwise_distances_indices(
+            self.embedding_, metric=self.metric_out, indices=self.neg_indices_
+        )[0]
+        D = self._a * D**self._b
+        return -sum_red(D.clamp(min=1e-12).log() - D.log1p(), dim=(0, 1))
 
     def _attractive_loss(self):
-        Q = self.affinity_out(self.embedding_, indices=self.NN_indices_)
-        Q = Q / (Q + 1)  # stabilization trick, PR #856 from UMAP repo
-        return cross_entropy_loss(self.affinity_in_, Q)
+        D = symmetric_pairwise_distances_indices(
+            self.embedding_, metric=self.metric_out, indices=self.NN_indices_
+        )[0]
+        D = self._a * D**self._b
+        return cross_entropy_loss(self.affinity_in_, -D.log1p(), log=True)
